Get all orders implementation

+++ db/products/products-utils.c
@@ -80,7 +80,7 @@ int count_all_category_products(char *cName){
 
     sqlite3_bind_text(stmt, 1, cName, -1, SQLITE_STATIC);
     if (sqlite3_step(stmt) == SQLITE_ROW){
-        count = atof((const char*)sqlite3_column_text(stmt, 0));
+        count = atoi((const char*)sqlite3_column_text(stmt, 0));
     }
 
     sqlite3_finalize(stmt);
